requirements.txt
Pillow==9.1.0
Django==3.2.13
django-ckeditor==6.4.0
django-environ==0.8.1
pymysql==1.0.2
django-currentuser==0.5.3
mysqlclient==2.1.0
